Move TOML calls into new reader interface
